Factor some functions into yield tools
